renderer_vulkan: Rework sychronization
* Turning on sync validation revealed the horrors...

* This commit therefore tightens image and buffer syncrhonization using the appropriate barriers. In addition the general layout is used everywhere as it simplifies layout tracking significantly and works fine for other emus like yuzu, skyline, cemu etc

+++ src/video_core/rasterizer_cache/utils.h
@@ -33,7 +33,7 @@ union ClearValue {
     Common::Vec4f color;
     struct {
         float depth;
-        u8 stencil;
+        u32 stencil;
     };
 };
